Revert "refactor(comp/checks/winregistry): migrate to V2 component pattern (#…"
This reverts commit db0a0c2bd8d1da67b214a5bc840ecbbd26d2174b.
