+++ comp/checks/winregistry/component.go
