--- comp/checks/winregistry/fx/fx.go
@@ -1,21 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2023-present Datadog, Inc.
-
-//go:build windows
-
-// Package fx provides the fx module for the winregistry component.
-package fx
-
-import (
-	winregistryimpl "github.com/DataDog/datadog-agent/comp/checks/winregistry/impl"
-	"github.com/DataDog/datadog-agent/pkg/util/fxutil"
-)
-
-// Module defines the fx options for this component.
-func Module() fxutil.Module {
-	return fxutil.Component(
-		fxutil.ProvideComponentConstructor(winregistryimpl.NewComponent),
-	)
-}
