--- comp/checks/winregistry/mock/mock.go
@@ -1,20 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2023-present Datadog, Inc.
-
-//go:build test && windows
-
-// Package mock provides a mock for the winregistry component.
-package mock
-
-import (
-	"testing"
-
-	winregistry "github.com/DataDog/datadog-agent/comp/checks/winregistry/def"
-)
-
-// Mock returns a mock for the winregistry component.
-func Mock(_ *testing.T) winregistry.Component {
-	return struct{}{}
-}
